+++ src/pages/Admin/Charity/Banking/PayoutMethods/DeletePrompt.tsx
