+++ src/pages/Admin/Charity/Banking/PayoutMethods/Table.tsx
