+++ src/pages/Admin/Charity/Banking/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Banking";
+export { default } from "./PayoutMethods";
