--- src/pages/Admin/Charity/PayoutMethods/List/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./List";
